add linux support and restructure source code.
